+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [ES_Utilities](https://github.com/ESpmdark/ES_Utilities)
 
-## Version: 2.0.2
-- Fixed a wrong reference in currencies init.
+## Version: 2.0.3
+- Cleaned up some text spacing in the vaultprogress display.
+- Fixed a case of nil data on when initiating affix info.
+- New module: ExtraActionButton art tweak
+- New module: LossOfControl art tweak